requirements.txt
fastapi
fastapi-code-generator
numpy
pandas
openai
pydantic
requests
httpx
python-dotenv
uvicorn
chromadb